Snap items registration is now scattered logically between sources
